+++ pyzo/core/splash.py
@@ -35,7 +35,7 @@ QLabel {
 splash_text = """
 <p>{text_title}</p>
 <p>{text_version} {version}</p>
-<p>{text_os} <a href='http://www.pyzo.org/'>http://pyzo.org</a></p>
+<p>{text_os} <a href='http://pyzo.org/'>http://pyzo.org</a></p>
 """